requirements.txt
google-cloud-bigquery==3.11.0